Some cleanup of GUIWindow.cpp
