+++ src/GUI/GUIWindow.cpp
@@ -56,7 +56,7 @@ GUIWindow *pPrimaryWindow;
 GUIWindow *pGUIWindow_CurrentMenu;
 GUIWindow *pDialogueWindow;
 GUIWindow_House *window_SpeakInHouse;
-GUIWindow_Scroll *pGUIWindow_ScrollWindow; // reading a message scroll
+GUIWindow_MessageScroll *pGUIWindow_ScrollWindow; // reading a message scroll
 GUIWindow *ptr_507BC8;  // screen 19 - not used?
 GUIWindow *pGUIWindow_CastTargetedSpell;
 GUIWindow *pGameOverWindow; // UIMSG_ShowGameOverWindow
@@ -504,10 +504,6 @@ void GUIWindow_BooksButtonOverlay::Update() {
     render->DrawTextureNew(uFrameY / 640.0f, uFrameX / 480.0f, pButton->vTextures[0]);
 }
 
-void GUIWindow_Scroll::Update() {
-    CreateScrollWindow();
-}
-
 void OnButtonClick::Update() {
     if (bPlaySound) {
         pAudioPlayer->playUISound(SOUND_StartMainChoice02);
@@ -644,49 +640,6 @@ void GUI_UpdateWindows() {
     }
 }
 
-void CreateScrollWindow() {
-    GUIWindow_Scroll &a1 = *pGUIWindow_ScrollWindow;
-    a1.sHint.clear();
-    a1.uFrameX = 1;
-    a1.uFrameY = 1;
-    a1.uFrameWidth = 468;
-    int textHeight = pFontSmallnum->CalcTextHeight(pMessageScrolls[pGUIWindow_ScrollWindow->scroll_type], a1.uFrameWidth, 0);
-    unsigned int v0 = textHeight + 2 * (unsigned char)pFontCreate->GetHeight() + 24;
-    a1.uFrameHeight = v0;
-    if ((v0 + a1.uFrameY) > 479) {
-        v0 = 479 - a1.uFrameY;
-        a1.uFrameHeight = 479 - a1.uFrameY;
-    }
-    a1.uFrameZ = a1.uFrameWidth + a1.uFrameX - 1;
-    a1.uFrameW = v0 + a1.uFrameY - 1;
-    a1.DrawMessageBox(0);
-    a1.uFrameX += 12;
-    a1.uFrameWidth -= 28;
-    a1.uFrameY += 12;
-    a1.uFrameHeight -= 12;
-    a1.uFrameZ = a1.uFrameWidth + a1.uFrameX - 1;
-    a1.uFrameW = a1.uFrameHeight + a1.uFrameY - 1;
-
-    const std::string &name = pItemTable->pItems[pGUIWindow_ScrollWindow->scroll_type].name;
-
-    a1.DrawTitleText(pFontCreate, 0, 0, Color(), fmt::format("{::}{}\f00000\n", colorTable.PaleCanary.tag(), name), 3);
-    a1.DrawText(pFontSmallnum, {1, pFontCreate->GetHeight() - 3}, Color(), pMessageScrolls[pGUIWindow_ScrollWindow->scroll_type], 0, 0, Color());
-}
-
-//----- (00467F48) --------------------------------------------------------
-void CreateMsgScrollWindow(ITEM_TYPE mscroll_id) {
-    if (!pGUIWindow_ScrollWindow && isMessageScroll(mscroll_id)) {
-        pGUIWindow_ScrollWindow = new GUIWindow_Scroll({0, 0}, render->GetRenderDimensions(), mscroll_id, "");
-    }
-}
-
-//----- (00467F9F) --------------------------------------------------------
-void free_book_subwindow() {
-    if (pGUIWindow_ScrollWindow) {
-        pGUIWindow_ScrollWindow->Release();
-        pGUIWindow_ScrollWindow = nullptr;
-    }
-}
 //----- (004226EF) --------------------------------------------------------
 void SetUserInterface(PartyAlignment align, bool bReplace) {
     extern void set_default_ui_skin();
